The sixth version
There is not much to say about this. Also this one is released
under GPLv3 or later. I forgot to mention that in the previous
commits but I guess that doesn't matter much. Anyway I don't
recommend using this or any of the previous versions as base for
new development but instead focus to the current development
version or version seven when it's released.